fix: fixed experience fields and added valid field
- spent total is now an u16 as the exp can be gt 900 in case the sheet is invalid
- remaining is now an i16 as the exp can be negative if the sheet is invalid
- added a valid field to the PlayerCharacter struct to make it easier to see if a sheet is valid (instead of having to look up and understand the exp values)

+++ src/character_db/experience_information.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 pub struct ExperienceInformation {
     pub start_value: u8,
-    pub spent_total: u8,
-    pub available: u8,
+    pub spent_total: u16,
+    pub available: i16,
     pub received_total: u8,
 }
